Fixed create and delete review services

+++ vaporAPI/Service/ReviewService.cs
@@ -29,6 +29,10 @@ namespace vaporAPI.Service
                 return new ApiResponse<ReviewDto>(null, false, "Manga with the given id does not exist", HttpStatusCode.NotFound);
 
             var review = createReviewDto.ToCreateFromDto(userId, manga);
+            var existingReview = manga.Reviews.Select(r => r.UserId).Contains(userId);
+
+            if (existingReview)
+                return new ApiResponse<ReviewDto>(null, false, "You have already reviewed this manga", HttpStatusCode.Conflict);
 
             manga.Reviews.Add(review);
             manga.AverageRating = (decimal?)manga.Reviews.Average(r => r.Rating);
@@ -38,11 +42,25 @@ namespace vaporAPI.Service
             return new ApiResponse<ReviewDto>(review.ToReviewDto(), true, "Review created successfully", HttpStatusCode.OK);
         }
 
-        public async Task<ApiResponse<ReviewDto>> DeleteReviewAsync(int id)
+        public async Task<ApiResponse<ReviewDto>> DeleteReviewAsync(int id, string userId)
         {
+            if (userId == null)
+                return new ApiResponse<ReviewDto>(null, false, "User with the given id does not exist", HttpStatusCode.NotFound);
+
             var review = await _reviewRepo.GetByIdAsync(id);
             if (review == null)
                 return new ApiResponse<ReviewDto>(null, false, "Review with the given id does not exist", HttpStatusCode.NotFound);
+
+            if (review.UserId != userId)
+                return new ApiResponse<ReviewDto>(null, false, "You are not authorized to delete this review", HttpStatusCode.Unauthorized);
+
+            var manga = await _reviewRepo.MangaExists(review.MangaId);
+            if (manga == null)
+                return new ApiResponse<ReviewDto>(null, false, "Manga with the given id does not exist", HttpStatusCode.NotFound);
+
+            manga.Reviews.Remove(review);
+            manga.AverageRating = manga.Reviews.Count > 0 ? (decimal?)manga.Reviews.Average(r => r.Rating) : 0;
+
             await _reviewRepo.DeleteAsync(review);
             return new ApiResponse<ReviewDto>(review.ToReviewDto(), true, "Successfully deleted a review", HttpStatusCode.OK);
         }
